refactor: use script to update options
